Extracted outcome list item header into separate component
